directory struct optimization
