+++ Assets/Scripts/Data Management/CombatSystem.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ed1dea9ac6391db459b81c6d173c779d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
