+++ Assets/Scripts/Data Management/CombatSystem/CombatMessenger.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: af2bfafa91e1c8a48b58df9929212d3a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
