+++ Assets/Scripts/Data Management/CombatSystem/CombatSystem.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: eb4bbe044c55cb047bc70dff526b0715
+guid: 6b4c1a4c943ccf64794ebdccb0850708
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
