+++ Assets/Scripts/Data Management/Crafting.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 48cf83408b3fbbd4ca330316f5284d0c
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
