+++ Assets/Scripts/Data Management/Crafting/Recipe.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2e7d6a707000d8a4296532233188ff7e
+guid: c5042a6f08ca51b4cb77534c62a3f658
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
