+++ Assets/Scripts/Data Management/Equipment.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 30080796c77258a44a6bffb4024b0b44
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
