+++ Assets/Scripts/Data Management/Equipment/Equipment.cs
@@ -1,0 +1,5 @@
+﻿
+abstract class Equipment : Item, IUsable
+{
+    public abstract bool Use();
+}
