+++ Assets/Scripts/Data Management/Equipment/Equipment.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bf4eff4f794ca214f836500005804c01
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
