+++ Assets/Scripts/Data Management/Equipment/EquipmentSlot.cs
@@ -1,0 +1,4 @@
+﻿class EquipmentSlot
+{
+
+}
