+++ Assets/Scripts/Data Management/Equipment/EquipmentSlot.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ddca12a36b0f88c49bec28c50d54da13
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
