+++ Assets/Scripts/Data Management/Equipment/Weapon.cs
@@ -1,0 +1,19 @@
+abstract class Weapon : Equipment
+{
+
+}
+
+abstract class Ammo : StackableItem
+{
+
+}
+
+abstract class Gun : Weapon
+{
+
+}
+
+abstract class MeleeWeapon : Weapon
+{
+
+}
