--- Assets/Scripts/Data Management/Weapon.cs
@@ -1,4 +1,0 @@
-abstract class Weapon : Item, IUsable
-{
-    public abstract bool Use();
-}
